requirement.txt
Beautifulsoup4 
requests
python-docx
openai
lxml
tqdm